Pong is a two player game where you must knock the “ball” to your opponent’s side and is one of the simplest internet games. The reason I would want to make a similar game is that it’s a pretty common starter game to make, similar to flappy bird where it has few assets. The interactions are simple arrow keys where you move up or down with either W and A or the arrow keys depending on which player you are, as well as the collisions where the ball is knocked back and forth. The win condition would be knocking the ball to the opposite side enough times, while the lost condition would be the opposite where you miss the ball enough times to lose.
Tetris is a pretty old game that has been remade multiple times. It is much more complex than pong, meaning it would have a bigger
challenge. The reason I would make a similar game is that it has a lot of assets that would be useful to learn about, as well as the
concept of collisions where certain items can knock into each other. The interaction requires moving the current tetris piece around
by translation and rotation, as well as switching out a previous piece and making sure a row is destroyed after being filled. In some
cases, the win condition would be getting a certain amount of points depending on how many rows you’ve destroyed. The lose condition
in all games is if you’ve run out of space to place blocks.

Lastly, Snake is in between Pong and Tetris in terms of complexity, at least in my opinion. This is the reason why I would choose it,
being a middle ground between the two games. The interactions would be a WASD or arrow key script that allows the snake to move around
and a condition that determines if you’ve eaten an apple. The win condition would be growing large enough to cover the entire map,
having no more space to move. The lose condition would be knocking into the wall or your own tail.